a/util: Make sink split support >2 downstreams
I needed this for the constellation tracker, as we have (in theory
unlimited) cameras running at once, and each camera gets it's own copy
of the frame. We hardcode a group of cameras of 5 right now.

This uses the same limit as the constellation tracker. This limit should
be increased whenever a usecase needs more than 5 downstreams.

Part-of: <https://gitlab.freedesktop.org/monado/monado/-/merge_requests/2498>