a/util: Add "reinterpret bayer format as l8" quirk
This is useful for when cameras expose raw bayer formats, and the on-device debayering does a bad job when the image is later converted to greyscale, such as for tracking fiducial markers, or doing camera calibration with a checkerboard.

This was added for the PlayStation Eye camera, which has really poor debayering, and doing this reinterpret trick helps get more acurate calibrations and fiducial tracking.

Part-of: <https://gitlab.freedesktop.org/monado/monado/-/merge_requests/2540>